Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -3,7 +3,7 @@
 int main() {
     float a,b,c;
     scanf("%f %f %f",&a,&b,&c);
-    int z=(a+b+c)/3;
+    float z=(a+b+c)/3;
     printf("Average: %.2f",z);
     return 0;
 }
